refactor: move grid finder related logic to `helpers/grid`
* move `GridPositionFinder` type to `grid.d.ts` file
